# api/openapi.yaml
openapi: 3.1.0
info:
  title: Le Hangar
  version: "{{VERSION}}"
  summary: A small cloud's control plane — ask for a machine, get one within your limits.
  description: |
    One API for everything a person can ask for. What can be asked for is
    whatever the enabled plugins declare: `GET /v1/types` is the catalogue,
    each type with the JSON Schema of its spec and its actions, and the
    command line and the console are drawn from it.

    **The ask.** Every request goes through the same questions, and every no
    says which one and why, with the numbers: signed in? (401) · in a tier?
    (403 `no-tier`) · the zone open to the tier and able to host the type?
    (403 `zone`, 422 `unavailable-here`) · the spec valid? (422 `schema`) ·
    does the plugin accept it? (422 `plugin-refused`) · within the tier's
    limits? (403 `limit`, e.g. "2 of 2 toy.boxes used; this asks for 1
    more") · room in the zone's pools? (409 `room`, e.g. "zone lab's
    borrowed room is held for priority (while guest 4100 runs): ask again
    when it ends, or for guaranteed room"). Then an **operation** is
    started and returned (202): poll it, or wait on it with `?wait=`.

    **Room.** A zone may count its memory in two pools — guaranteed (booked
    while a resource lives) and spot (lent while it runs, taken back when a
    reservation needs it). A reservation waiting on a guest is claimed by
    that guest's hook before it starts and released after it stops
...

...

...

...

...

...

...
  license:
...
servers:
...
security:
  - bearer: []
...
paths:
...
  /v1/resources/{id}/streams/{stream}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
      - { name: stream, in: path, required: true, schema: { type: string } }
      - { name: cols, in: query, schema: { type: integer, minimum: 1, maximum: 1000 }, description: "The window as it opens, in characters." }
      - { name: rows, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
    get:
      tags: [resources]
      summary: One of the resource's streams, held open — a machine's terminal. A WebSocket.
      description: |
        The one call that is not a request and its answer: a WebSocket
        (RFC 6455), asked with `Upgrade: websocket` and the same bearer token
        as everything else. A type's streams are in the catalogue
        (`GET /v1/types`, `streams`).

        **Its owner alone.** A stream reaches inside a resource: an operator
        sees a machine, stops it, deletes it, and is refused here (403
        `owner`), as is someone it is shared with; a read-only token is
        refused (403 `scope`). **One at a time** per resource and stream: a
        second opening ends the first, which is told (close code 4001).

        A refusal is an ordinary problem document, before anything is
        switched — 409 `engine` for a machine that is stopped. Once open: a
        binary message from the client is what is typed, a text message
        `{"size":[cols,rows]}` a changed window; a binary message from the
        brain is what the resource says. It ends with a close frame from the
        brain: 4000 and why in words (« the machine was stopped »), 4001
        when its owner opened it elsewhere.

        **Held on its credential.** What this call asked is asked again
        every half minute while the stream is open: a token revoked or
        ended, a person no longer in any tier, closes it (4000, with the
        reason). A client whose token is short-lived sends the next one
        before that, as a text message `{"token":"<the new bearer>"}` — the
        same person's, or the stream is closed.

        Two audit lines: the opening, and the closing — how long, how many
        bytes each way, why. Never what passed.
      responses:
        "101":
          description: Open — the connection is a WebSocket from here.
        "426":
          description: "Asked without `Upgrade: websocket`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A token signed by the operator's OIDC provider for the configured client id (an ID token, or an access token in JWT form), or an API token (`hgr_…`).
  responses:
    Problem:
      description: Refused — the problem says which question failed, and why.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Id:
      type: string
      description: A short prefix naming the kind, a dash, seventeen hex digits.
      pattern: "^[a-z][a-z0-9]{0,7}-[0-9a-f]{17}$"
      examples: ["box-0123456789abcdef0", "op-0123456789abcdef0"]
...
    Problem:
      type: object
      required: [type, title, status, detail, kind]
      properties:
        type: { type: string, examples: ["urn:hangar:problem:limit"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, examples: ["2 of 2 toy.boxes used; this asks for 1 more"] }
        kind:
          type: string
          enum: [sign-in, no-tier, scope, not-found, bad-request, zone, unavailable-here, schema, plugin-refused, limit, room, busy, conflict, attached, members, shared, engine, plugin-down, internal]
        violations:
          type: array
          description: |
            For `schema` and `plugin-refused` — each field that does not fit.
            And for a name or a description that is refused (`bad-request`,
            or `conflict` for a name already held): the field is then `name`
            or `description`, the resource's own — no pointer into its spec.
          items:
            type: object
            properties:
              field: { type: string, description: "A JSON pointer into the spec or the params (/cores) — or name, description." }
              reason: { type: string }
        refusals:
          type: array
          description: |
            For `limit` — every dimension the request exceeds, with the
            numbers; or the meter whose month is spent ("10.1 of 10
            vCPU-hours (machines.vcpu_hours) used in October 2026: it is back
            on 1 November").
          items:
            type: object
            required: [reason, dimension, message]
            properties:
              reason: { type: string, enum: [limit, choice, meter] }
              dimension: { type: string }
              limit: { type: integer }
              used: { type: integer }
              asked: { type: integer }
              value: { type: string }
              allowed: { type: array, items: { type: string } }
              consumed: { type: number, description: For `meter` — what the month has consumed. }
              period: { type: string, description: For `meter` — the month. }
              resets: { type: string, format: date-time, description: For `meter` — when the next month begins. }
              message: { type: string }
        room:
          type: object
          description: For `room` — the pool the request does not fit, with the numbers (MiB).
          required: [pool, zone, message]
          properties:
            pool: { type: string, enum: [guaranteed, spot] }
            zone: { type: string }
            held_for: { type: string, description: The reservation holding the borrowed room, when that is why. }
            pool_mb: { type: integer }
            used_mb: { type: integer }
            asked_mb: { type: integer }
            message: { type: string }
        operation:
          $ref: "#/components/schemas/Id"
          description: For `busy` — the operation holding the resource.